backend/requirements.txt
fastapi[all]
uvicorn[standard]
langchain
langgraph
langchain-ollama
fastmcp
sqlalchemy
asyncpg
python-multipart
edge-tts
faster-whisper
python-dotenv
websockets
langchain_mcp_adapters
sentence-transformers
pgvector
psycopg2-binary
pydantic-settings
numpy
passlib[bcrypt]
python-jose[cryptography]